VS C++20 compilation

+++ src/LibSL/Mesh/VertexFormat_reference.h
@@ -149,7 +149,7 @@ namespace LibSL {
       enum { stop = MVF_is_item<Head,Item>::value     };
       enum { next = MVF_offset_item<Tail,Item>::value };
     public:
-      enum { value = (stop == 0) ? ((next == -1) ? -1 : (Head::size_of + next)) : 0 };
+      enum { value = (stop == 0) ? ((next == -1) ? -1 : (static_cast<int>(Head::size_of) + static_cast<int>(next))) : 0 };
     };
 
     // -------------------------------------------------
